Fix mock_postgresql_db spec & create conftest
