+++ src/students/alternate_emails/router.py
@@ -26,4 +26,3 @@ def modify_alternate_emails(
     except Exception as e:
         db.rollback()
         raise HTTPException(status_code=500, detail=f"Unexpected error: {str(e)}")
-
