+++ tests/test_main.py
@@ -1,5 +1,4 @@
 from datetime import datetime, timezone
-import os
 import bson
 from pydantic import ValidationError
 import pymongo
@@ -10,17 +9,11 @@ import pymongo.mongo_client
 import pytest
 import mongomock
 from fastapi.testclient import TestClient
-from src.config import ACCELERATE_FLEX_COLLECTION, APPLICATIONS_COLLECTION, COURSES_COLLECTION, MONGO_DATABASE_NAME, PATHWAY_GOALS_COLLECTION
-from src.database.mongo.core import get_mongo
-from src.database.mongo.services import init_collections
+from src.config import ACCELERATE_FLEX_COLLECTION, APPLICATIONS_COLLECTION, COURSES_COLLECTION, PATHWAY_GOALS_COLLECTION
 from src.database.mongo.tmp_schemas import AccelerateFlexBase, ApplicationModel, CourseBase, DeepWork, PathwayGoalBase
-from src.database.postgres.core import make_session
 from src.database.postgres.models import Student, StudentEmail
 from src.main import app
-from sqlalchemy.orm import Session
-from unittest.mock import MagicMock
 from sqlalchemy.exc import SQLAlchemyError
-
 
 client = TestClient(app)
 
@@ -33,40 +26,6 @@ def test_confirm_conn():
     response = client.get("/test-connection")
     assert response.status_code == 200
     assert response.json() == {"message": "Database connection succeeded"}
-
-@pytest.fixture(scope="function")
-def mock_mongo_db():
-    """Injection for MongoDB dependency intended for fast, in-memory unit testing"""
-    mock_client = mongomock.MongoClient()
-    db = mock_client[MONGO_DATABASE_NAME]
-
-    # Apply indexes (json schema validators cannot be enforced in mongomock)
-    init_collections(db, with_validators=False)
-
-    # Override FastAPI's database dependency
-    app.dependency_overrides[get_mongo] = lambda: db
-
-    yield db # Provide the mock DB instance
-
-    app.dependency_overrides.pop(get_mongo) # Clean up override(s) after test
-    mock_client.close()
-
-@pytest.fixture(scope="function")
-def real_mongo_db():
-    """Injection for MongoDB dependency intended for wide scope, accurate integration testing"""
-    # Consider replacing this with a conditionally created local instance (in GitHub Actions)
-    client = pymongo.MongoClient(os.environ.get("CTI_MONGO_URL"))
-    test_db_name = "test_" + MONGO_DATABASE_NAME
-    db = client[test_db_name]
-    init_collections(db, with_validators=True)
-
-    app.dependency_overrides[get_mongo] = lambda: db
-
-    yield db
-
-    app.dependency_overrides.pop(get_mongo)
-    client.drop_database(db)
-    client.close()
 
 class TestCreateApplication:
     def test_success_min_required_fields(self, mock_mongo_db: mongomock.Database):
@@ -461,16 +420,6 @@ class TestCourses:
         assert found_course is not None
         assert found_course["course_id"] == course_id
         assert prev_count + 1 == courses_collection.count_documents({})
-
-
-@pytest.fixture(scope="function")
-def mock_postgresql_db():
-    """Fixture to mock a PostgreSQL database session for testing."""
-    db = MagicMock(spec=Session)
-    app.dependency_overrides[make_session] = lambda: db
-    yield db
-    app.dependency_overrides.pop(make_session)
-
 
 class TestModifyAlternateEmails:
